+++ loadCsv.py
@@ -4,7 +4,7 @@ from knnPure import KnnClasifier
 from sklearn.model_selection import train_test_split
 
 import numpy as np
-import pruebaMeshgrid as msg
+import plotter as msg
 
 df = pd.read_csv('datasets-short.csv')
